refactor(backend): move task create into utils
Extract task integration ops (create + OAuth helpers) from routers to
utils/task_integrations_ops.py so task_sync no longer imports routers.
Add import-boundary ratchet and Todoist characterization tests.

Verification: pytest tests/unit/test_utils_import_boundary.py
tests/unit/test_task_integrations_ops.py tests/unit/test_task_sync_async.py
